requirements.txt
streamlit
firebase-admin
google-generativeai
reportlab
